refactor: address review comments - move providers directory and improve debug output
Review fixes:
- Move providers directory from eval/providers/ to commands/providers/
  * Better organization and consistency with flattened CLI structure
  * Update all imports across codebase
- Only show debug messages when --debug flag is enabled
  * Letta Folder ID, Collection ID, Conversation ID now respect debug flag
  * Cleaner output in normal mode
- Collect all missing Albert provider env vars and report together
  * Better UX - users see all issues at once instead of one at a time
- Update docstrings from "init command" to "setup command"
  * Consistency with renamed setup command

Files changed:
- Moved: apps/cli/src/cli/commands/eval/providers/ → apps/cli/src/cli/commands/providers/
- Updated: apps/cli/src/cli/commands/generate_dataset.py
- Updated: apps/cli/tests/test_setup.py, test_generate_dataset.py, test_document_preprocessor.py

Test status: 70/71 tests passing (1 pre-existing env isolation test failure)

Addresses all review comments from PR #44.
