+++ apps/cli/src/cli/commands/providers/schema.py
